add test

+++ vt_taskc.py
@@ -154,6 +154,8 @@ def capa_extraction(path_file,path_rules='/mnt/data/capa-rules-5.1.0/',path_sign
         name_file = os.path.basename(path_file)
         path_dir = '/mnt/pst/jsons_capa'
         path_file_json = '%s.capa' % os.path.join(path_dir, name_file)
+        if os.path.isfile(path_file_json):
+            return True
         try:
             fw = open(path_file_json, 'w')
             if fw:
